HPCC-32689 Include/Exclude pending scopes in metrics
