+++ esp/src/src-react/components/ECLArchive.tsx
@@ -39,7 +39,7 @@ export const ECLArchive: React.FunctionComponent<ECLArchiveProps> = ({
     const [fullscreen, setFullscreen] = React.useState<boolean>(false);
     const [dockpanel, setDockpanel] = React.useState<ResetableDockPanel>();
     const [_archiveXmlStr, _workunit2, _state2, archive, refreshArchive] = useWorkunitArchive(wuid);
-    const [metrics, _columns, _activities, _properties, _measures, _scopeTypes, _fetchStatus, refreshMetrics] = useWorkunitMetrics(wuid, scopeFilterDefault, nestedFilterDefault);
+    const { metrics, refresh: refreshMetrics } = useWorkunitMetrics(wuid, scopeFilterDefault, nestedFilterDefault);
     const [markers, setMarkers] = React.useState<{ lineNum: number, label: string }[]>([]);
     const [selectionText, setSelectionText] = React.useState<string>("");
     const [selectedMetrics, setSelectedMetrics] = React.useState<IScope[]>([]);
