+++ esp/src/src/nls/hpcc.ts
@@ -411,6 +411,7 @@ export = {
         IgnoreGlobalStoreOutEdges: "Ignore Global Store Out Edges",
         Import: "Import",
         Inactive: "Inactive",
+        IncludePendingItems: "Include pending items",
         IncludeSlaveLogs: "Include worker logs",
         IncludeSubFileInfo: "Include sub file info?",
         Index: "Index",
